# packages/server-monitor/swagger_codegenerator/SwaggerDefintion.json
{"swagger":"2.0","info":{"description":"Mail server","version":"1.0","title":"Mail server gateway"},"basePath":"/","tags":[{"name":"Mail Server Gateway","description":""}],"schemes":["http"],"paths":{"/management/SendMail":{"put":{"summary":"Send an e-mail to test mail server gateway","description":"","operationId":"SendTestMail","parameters":[{"name":"MailData","required":true,"in":"body","schema":{"$ref":"#/definitions/MailData"}},{"type":"boolean","name":"useKafka","required":true,"in":"query"}],"responses":{"200":{"description":"","schema":{"type":"string"}}},"tags":["management"],"produces":["application/json"],"consumes":["application/json"]}},"/management/GetStatus":{"get":{"summary":"Get status","description":"Get status of server","operationId":"GetStatus","responses":{"200":{"description":"Returns the status of server","schema":{"$ref":"#/definitions/Statusresult"}}},"tags":["management"],"produces":["application/json"],"consumes":["application/json"]}},"/management/MailAccounts":{"get":{"summary":"Get mail accounts","description":"Get mail accounts","operationId":"MailAccounts","responses":{"200":{"description":"Returns the mail accounts","schema":{"$ref":"#/definitions/MailAccountsResultImpl"}}},"tags":["management"],"produces":["application/json"],"consumes":["application/json"]}},"/management/AddMailAccount":{"post":{"summary":"Add mail account","description":"Add mail account to mail server (e.g. \"user@world.com\") ","operationId":"AddAccount","parameters":[{"type":"string","name":"password","required":true,"in":"query","description":"The mail account password"},{"type":"string","name":"account","required":true,"in":"query","description":"The mail account"}],"responses":{"200":{"description":"","schema":{"$ref":"#/definitions/AddMailAccountResultImpl"}}},"tags":["management"],"produces":["application/json"],"consumes":["application/json"]}},"/management/DeleteMailAccount":{"post":{"summary":"Delete mail account","description":"Delete mail account to mail server \"user@world.com\" ","operationId":"DeleteAccount","parameters":[{"type":"string","name":"account","required":true,"in":"query","description":"The mail account"}],"responses":{"200":{"description":"","schema":{"$ref":"#/definitions/DeleteMailAccountResultImpl"}}},"tags":["management"],"produces":["application/json"],"consumes":["application/json"]}},"/management/Reset":{"post":{"summary":"Reset","description":"Reset database","operationId":"Reset","responses":{"200":{"description":"","schema":{"$ref":"#/definitions/ResetResultImpl"}}},"tags":["management"],"produces":["application/json"],"consumes":["application/json"]}}},"definitions":{"MailData":{"type":"object","properties":{"From":{"type":"string"},"To":{"type":"string"},"Subject":{"type":"string"},"Content":{"type":"string"}},"required":["From","To","Subject","Content"]},"Statusresult":{"type":"object","properties":{"StatusAsJson":{"type":"string"}},"required":["StatusAsJson"]},"MailAccountsResultImpl":{"type":"object","properties":{"Accounts":{"type":"array","items":{"type":"string"}}},"required":["Accounts"]},"AddMailAccountResultImpl":{"type":"object","properties":{"Msg":{"type":"string"}},"required":["Msg"]},"DeleteMailAccountResultImpl":{"type":"object","properties":{"Msg":{"type":"string"}},"required":["Msg"]},"ResetResultImpl":{"type":"object","properties":{"Msg":{"type":"string"}},"required":["Msg"]}}}